Added A* Pathfinding For Bat Enemy and Bat Flying Animation
Imported Files for A*. This allows us to have a flying/grounded enemy to chase our player. Not sure if we will end up using this but it is a start to AI. Added the bat flying animation to make it look better also. Check it out in my test scene.